@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "Flask-Reuploaded"
-version = "1.4.1.dev"
+version = "1.5.0.dev"
 description = "Flexible and efficient upload handling for Flask"
 license = {text = "MIT"}
 authors = [
@@ -13,7 +13,7 @@ authors = [
 maintainers = [
     {name = "Jürgen Gmach", email = "juergen.gmach@googlemail.com"},
 ]
-requires-python = ">= 3.8"
+requires-python = ">= 3.10"
 dynamic = ["readme"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -22,11 +22,10 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
     "Topic :: Software Development :: Libraries :: Python Modules",
     "Framework :: Flask",
